src/requirements.txt
# Python dependencies - install with: pip install -r requirements.txt
#
# System-level binary tool dependencies (NOT pip packages):
#   - nmap      : 全端口扫描与服务识别（LLM 驱动的端口发现核心依赖）
#   - netcat    : 网络连接与端口探测（可选）
#   - xray      : Web 漏洞自动化扫描器（Scan Agent 核心依赖）
#   - curl      : HTTP 请求工具
#   - wget      : 文件下载工具
#   - docker    : 漏洞靶机容器运行环境（Docker Desktop for Windows）
#
# Windows 一键安装所有系统工具（以管理员身份运行 PowerShell）:
#   Set-ExecutionPolicy -Scope Process -ExecutionPolicy Bypass
#   .\install_tools.ps1

beautifulsoup4==4.12.3
ipython==8.12.3
jsonlines==4.0.0
langchain==0.2.15
langchain_community==0.2.15
langchain_core==0.2.37
langchain_nvidia_ai_endpoints==0.2.2
langchain_openai==0.1.23
langgraph==0.2.16
nest_asyncio==1.6.0
paramiko==3.4.0
playwright==1.57.0
PyYAML==6.0.1
Requests==2.32.3
termcolor==2.4.0
